# api/openapi.yaml
openapi: 3.0.3
info:
  title: Micro-Blog API
  version: 0.1.0
  description: |
    Authenticated markdown micro-blog with a built-in URL shortener.
    Dual-token JWT (15min access + 7d refresh-cookie); RBAC enforced by a
    centralized permission/scope matrix.
servers:
  - url: http://localhost:8090
    description: Local dev (PORT from .env)
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /auth/login or /auth/refresh.
    RefreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly refresh token cookie issued by /auth/login.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid_input
          description: Machine-readable error code.
        message:
          type: string
          example: validation failed
          description: Human-readable detail (optional).
        fields:
          type: object
          additionalProperties:
            type: string
          example:
            username: must be at least 3 characters
            email: is not a valid email
          description: |
            Per-field validation messages, present only on `invalid_input`
            responses where one or more fields failed validation. Each value
            is a single human-readable message for that field.
    User:
      type: object
      required: [id, username, email, role_id, role]
      properties:
        id: { type: integer, format: int64, example: 1 }
        username: { type: string, example: alice }
        email: { type: string, format: email, example: alice@example.com }
        role_id: { type: integer, format: int64, example: 3 }
        role: { type: string, example: Author }
...
    TokenResponse:
      type: object
      required: [access_token, user]
      properties:
        access_token:
          type: string
          description: 'JWT access token; pass as `Authorization: Bearer <token>`.'
        user: { $ref: '#/components/schemas/User' }
...
    MeUpdateRequest:
      type: object
      description: |
        Partial profile update. `role_id` is intentionally **not** a property —
        even if a client sends it the server ignores it (no privilege escalation
        via /api/me).
      properties:
        username: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 8 }
  responses:
    BadRequest:
      description: Bad request — invalid body or input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Authentication required or token invalid/expired.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Conflict:
      description: Uniqueness constraint violation (username/email).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
paths:
...
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refresh
      x-roles: [anonymous, Subscriber, Author, Editor, Admin]
      summary: Rotate the refresh cookie and issue a new access token.
      description: |
        Reads the refresh_token cookie, deletes the old row server-side,
        inserts a new one, and sets a fresh cookie. Replaying the previous
        cookie returns 401.
      security:
        - RefreshCookie: []
      responses:
        '200':
          description: New access token; new refresh cookie set.
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/me:
    get:
      tags: [profile]
      operationId: getMe
      x-roles: [Subscriber, Author, Editor, Admin]
      summary: Read the authenticated user's profile.
      responses:
        '200':
          description: The caller's user row.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [profile]
      operationId: updateMe
      x-roles: [Subscriber, Author, Editor, Admin]
      summary: Change own username / email / password.
      description: |
        `role_id` is silently ignored if present in the body — self-service
        cannot escalate or demote.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MeUpdateRequest' }
      responses:
        '200':
          description: Updated profile.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
...